Create generic transactional withVerifiableLog method in PostgresDriver so that proof logic does not need to exist in this class
